Replace continue/start methods
Removed start_parsing and continue_parsing in favour of parse().
Move some continuous_parser tests to simplex and barrier classes.
