+++ src/grblogtools/parsers/util.py
@@ -37,11 +37,7 @@ def parse_lines(parser, loglines: Iterable[str]):
     """
     lines = iter(loglines)
     for line in lines:
-        if parser.start_parsing(line):
-            # Once the parser indicates start, use the parser to parse all
-            # remaining lines.
-            for line in lines:
-                parser.continue_parsing(line)
+        parser.parse(line)
 
 
 def parse_block(parser, log):
