+++ tests/test_cuts.py
@@ -1,4 +1,4 @@
-from helpers import *
+from .helpers import read_single
 
 
 def test_cuts():
